# Optimization for first time to read a random access readbuffer in hive