--- a/docs/zh/faq/general/who-is-using-clickhouse.md.bak
+++ b/docs/zh/faq/general/who-is-using-clickhouse.md.bak
@@ -1,1 +1,0 @@
-../../../en/faq/general/who-is-using-clickhouse.md