--- a/src/AggregateFunctions/AggregateFunctionArgMinMax.h
+++ b/src/AggregateFunctions/AggregateFunctionArgMinMax.h
@@ -16,7 +16,7 @@ namespace ErrorCodes
 }
 
 
-/// For possible values for template parameters, see AggregateFunctionMinMaxAny.h
+/// For possible values for template parameters, see 'AggregateFunctionMinMaxAny.h'.
 template <typename ResultData, typename ValueData>
 struct AggregateFunctionArgMinMaxData
 {