--- a/src/Client/ClientBase.h
+++ b/src/Client/ClientBase.h
@@ -108,7 +108,7 @@ protected:
 
 private:
     void receiveResult(ASTPtr parsed_query);
-    bool receiveAndProcessPacket(ASTPtr parsed_query, bool cancelled);
+    bool receiveAndProcessPacket(ASTPtr parsed_query, bool cancelled_);
     void receiveLogs(ASTPtr parsed_query);
     bool receiveSampleBlock(Block & out, ColumnsDescription & columns_description, ASTPtr parsed_query);
     bool receiveEndOfQuery();
@@ -259,6 +259,8 @@ protected:
     };
 
     std::vector<HostAndPort> hosts_and_ports{};
+
+    bool cancelled = false;
 };
 
 }