--- a/src/Storages/Cache/ExternalDataSourceCache.h
+++ b/src/Storages/Cache/ExternalDataSourceCache.h
@@ -34,10 +34,13 @@ class LocalFileHolder
 {
 public:
     explicit LocalFileHolder(RemoteFileCacheType::MappedHolderPtr cache_controller);
-    ~LocalFileHolder() = default;
+    explicit LocalFileHolder(RemoteFileCacheType::MappedHolderPtr cache_controller, std::unique_ptr<ReadBuffer> original_readbuffer_, BackgroundSchedulePool * thread_pool_);
+    ~LocalFileHolder();
 
     RemoteFileCacheType::MappedHolderPtr file_cache_controller;
     std::unique_ptr<ReadBufferFromFileBase> file_buffer;
+    std::unique_ptr<ReadBuffer> original_readbuffer;
+    BackgroundSchedulePool * thread_pool;
 };
 
 class RemoteReadBuffer : public BufferWithOwnMemory<SeekableReadBufferWithSize>
@@ -45,7 +48,7 @@ class RemoteReadBuffer : public BufferWithOwnMemory<SeekableReadBufferWithSize>
 public:
     explicit RemoteReadBuffer(size_t buff_size);
     ~RemoteReadBuffer() override = default;
-    static std::unique_ptr<ReadBuffer> create(ContextPtr context, IRemoteFileMetadataPtr remote_file_metadata, std::unique_ptr<ReadBuffer> read_buffer, size_t buff_size);
+    static std::unique_ptr<ReadBuffer> create(ContextPtr context, IRemoteFileMetadataPtr remote_file_metadata, std::unique_ptr<ReadBuffer> read_buffer, size_t buff_size, bool is_random_accessed = false);
 
     bool nextImpl() override;
     off_t seek(off_t off, int whence) override;
@@ -70,7 +73,8 @@ public:
     inline bool isInitialized() const { return initialized; }
 
     std::pair<std::unique_ptr<LocalFileHolder>, std::unique_ptr<ReadBuffer>>
-    createReader(ContextPtr context, IRemoteFileMetadataPtr remote_file_metadata, std::unique_ptr<ReadBuffer> & read_buffer);
+    createReader(ContextPtr context, IRemoteFileMetadataPtr remote_file_metadata, std::unique_ptr<ReadBuffer> & read_buffer, bool is_random_accessed);
+
 
     void updateTotalSize(size_t size) { total_size += size; }