Changelog category (leave one):
- Improvement



Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

For random access readbuffer in hive, the first time to read the readbuffer would use the original readbuffer instead of local file. When we read a parquet/orc format file, the readbuffer seeks to the end of the file, which will be blocked until the local file finishes download, and make the whold process slow.

In this optimization, the download process is executed when the original readbuffer is finished read.

running the following query
```
select day, count(*) from hive_test_table where day = '2022-02-20' and hour = '00' group by day
```

For reading parque files which are not cached in local disk, without this optimization
```
┌─day────────┬──count()─┐
│ 2022-02-20 │ 19773257 │
└────────────┴──────────┘

1 rows in set. Elapsed: 15.441 sec. Processed 19.77 million rows, 4.66 GB (1.28 million rows/s., 301.79 MB/s.)
```

with this optimization
```
┌─day────────┬──count()─┐
│ 2022-02-20 │ 19773257 │
└────────────┴──────────┘

1 rows in set. Elapsed: 5.905 sec. Processed 19.77 million rows, 4.66 GB (3.35 million rows/s., 789.14 MB/s.)
```

compare to local cached files reading
```
┌─day────────┬──count()─┐
│ 2022-02-20 │ 19773257 │
└────────────┴──────────┘

1 rows in set. Elapsed: 4.553 sec. Processed 19.77 million rows, 4.66 GB (4.34 million rows/s., 1.02 GB/s.)
```

> Information about CI checks: https://clickhouse.tech/docs/en/development/continuous-integration/